WIP: Quell valgrind errors with extent type.
The extent type field used as a u64 union, which depending on the
situation, may include undefined bits.  Valgrind understandably throws
an error when we call ffs() on this field, or test it against 0.

In this case, we really just wanted to check the type field itself, so
constrain the test to the first byte to make valgrind happy.

Question: the test against 0 is against the whole type, is this right?!

TODO: Port this to the kernel side.

Signed-off-by: Justin Husted <sigstop@gmail.com>